Implement adding new books to the library.
This was step 4 of the project, and it involved creating a dialog to pop up with a form for users to input new book information.
This involved adding event listeners to display the modal, change the read checkbox, as well as submit and process the form.
event.PreventDefault() was used on the submit event listener for the form so that the form would not try to send the data to the server.
Rather, I simply used the event's target to extract the user provided information from the form, create a new Book object with that data, and add it to the library array.

At first, I ran into some issues with the behavior of event.PreventDefault(), as it was not checking the validity of the form.
This was because I first had the event listener on the button click itself, rather than the form submit as a whole.