Add removing book functionality
This step involved adding a remove button to the book card creation process.
An event listener is added to each new remove button which extracts the index attrbute value from the DOM button element,
uses that index to find the correct book card element using the nth-child selector, and then removes that element from the DOM as well as the myLibrary array.