Implement toggle read button and logic
This part of the project required me to add a button to the book cards which allows users to change the state of the read status.
This involved adding a toggleReadStatus function in the Book constructor.
I also fixed a bug regarding the removal of book cards. I forgot to convert the index value to a Number, so performing addition actually ended up concatenating into a string rather than adding to a number.

This was the last step in Project: Library, which was a great applied example of using JavaScript objects and object constructors.